Create Wrongful death help
My husband died active duty in 1987. I was just informed within last few years. I was requested to file a claim. Because the truth was hidden from me on purpose for over 32 years. I finally found someone who knew what to do. So I filed a form 95. There is no statute of limitations since the military took blame but hid it from me. I now have the full report. And it took forever to find the right department and what to file.